requirements.txt
requests
Pillow
natsort
